Split Internal::Dimension into separate .h/.cpp file
No functional changes, just code motion to make an upcoming PR easier to review. (Splitting simplifies some order-of-#include problems in the upcoming PR.)
